Python/kookmin_django/foo/requirements.txt
chartkick==0.5.0
Django==1.10.3
django-bootstrap3==7.1.0
django-chartit==0.2.7
django-chartjs==1.1
django-crispy-forms==1.6.1
ez-setup==0.9
mysqlclient==1.3.7
rpy2==2.8.4
six==1.10.0
wfastcgi==2.2.1
